Made the refresh smarter
Dont use  location.reload(true)  if this was the result of a post
as that will do the action again.
Use   document.location=document.location  instead to go a GET

However if the current display was already a GET, then use location.reload(true)
so that the browser keeps your current location in the page.

Signed-off-by: David Deaves <HSV_Timing@dd.id.au>