Deal with websocket failure
Better handle the case where the web socket cannot be opened.

Start with a simple 70s refresh timeout.
Once the websocket is opened, cancel that timer, and add a handler on
close (often driven by an idle timeout of 60s)

Previously, when the socket failed to open, the 'onclose' handler was
being called.  This meant that when my web proxy could not properly
handle passing the websocket request, that the page was refreshing
as fast as the websocket failure could manifest.