Fix event update
Found event changes were not detected in timing.py
Changed  'event'  from a global, declaired after new_car()
to new_car.event to make sure every section is using the same
variable.  Could have used the 'global' marker, but followed what
I was doing for run_num, curr_car ....

With the new convention that -ve car numbers are used for out of order
events for the +ve version on that car, changed default curr_car to 1000
and next_missing_car to 1001

Changed start debounce back to 0.2s, as those event are now hidden with
the -ve car number, so keeping them is not bad.

Added a new_car.tick_val to track when tick_val wraps before start or
finish.  When writing start and finish, to the database, unwrap those
values by adding 4294967  (2 ^ 32 / 1000)