Handle single quotes in text fields
Using escapeString  is not correct when using bindValue, it adds an
extra single quote that makes sense if building a string to run, but not
when using bindValue.

Also single quotes in the name also confuses the javascript.  Rather
than trying to put original text in the javascript, create a hidden
field with the original text to compare to.