Better deal with red lights.
The  Start_times  panel mostly waits on the start_time web socket.  So
when someone red lights, it will refresh when they break the beam
(actually un-brake as they roll forward), but there is no green time yet
so no reasonable delta to display.  Often their rear wheels will pass
after green, so by luck we would get a refresh.  But this change tells
the page to also wait on the green_time web socket if there is a start
time without a matching green time in the first 3 lines.