Updates to  timing.py
Change the debounce from 0.2s to 0.3s

Some comments about the use/values of  new_car.state
+ #  0 - No Car (prev car finished) ,   1 - Car has green light  ,   2 - Car has a start time

Change  new_car.next  when no car is in the run queue from  99  to  666.

Add a few spaces to the end of the car number when writing to the state
files for each of the state files, making them nicer to read.

When we see a start before green with no next car, add the possible
false start record to this run, and the next run.
  (new_car.state == 0) and (new_car.next == 666)
This will happen for the first car of each new run.