Added button signal, and option to match car num
Added a button event  300ms  before the green signal.
This allows correct detection of a red light condition.

Also allowed the option to use the car_num from the green
record to pre-populate the  next_car  table so that we
can get a perfect match to replayed data.  Particularly
useful for the drag stuff where cars don't run in order.

Changed from using temporary files to hold the times for each of the
green, start and finish times to using process redirection

Exit on the end of finish entries.

When reducing the delta to allow for my processing speed,
better handle deltas less than the offset.