Make the hillclimb.php page show a full field
Don't just show cars with runs.

Modify the main query to be based around the  entrant_info  page
with LEFT JOINs into  class_info, results and et_order
Then change the sorting so that NULLS are last for timing.
Also don't try to display placing or times if there is no
result information.
NOTE:  There is some redundant checking in the display code

Similarly for the special placings.
Reverese the join to be a  entrant_info LEFT JOINed to et_order.
Then store the special place as empty if no result yet.