Add dragging to re-order run list.
Added drag functionality to the running_order panel.
You can click on the table cell with the car number, and then
drag it to where you want it to be in the list.
The selected car will end up in that spot it is dropped.
If dragged down, the the cars it passed will move up the list,
so it will be after the car it was dropped on.
If dragged up the list, the cars it passed will move down the list,
so it will be before the car it was dropped on.

To avoid violating the UNIQUE order number requirement on the
next_car table.  First, all the intermediate cars are moved
to past the end of the order with a +/-1 depending on direction
resuired.  Then the target car is given the desired order number,
then all the intermediate cars are moved back into the proper list
space.  This is all done in a  BEGIN -- COMMIT/ROLLBACK  transaction.

Also fixed an issue with the move to P3 queries.  The initial
query moving P1 and P2 up one spot, could not force the order to
make sure that P1 was done first, then P2.  So split this into
separate queries, again in a BEGIN -- COMMIT/ROLLBACK  transaction.
Because we were already being called with the  order  value of P2
(where our moving car was to end up, with current P1 and P2 to get lower
order numbers) we could select P1 as <ord, and P2 as ==ord.

NOTES:
The dragoverHandler() is tring to only show targets other than
the start as valid drop points.  This seems to work with Firefox,
but not with chrome.  The same test is performed in dropHandler
and effectively prevents a POST for when dropped at the start location.