Fix up input line setup
Pull the  set_mode(pigpio.INPUT)  and  set_pull_up_down  out
of the if block.
They are always inputs, and as the gpio is on the transistor side
of the opto isolator, the pull-up pull-down is not dependant on
the polarity of the input to the optoisolator.  Given there is a
physical pull-down.

Put the gpio callbacks and their rising or falling edge trigggering
into an  if/elif/else  block.

Will add command line flags or a config file in future.