Add Requisite to hostapd@.service
The  hostapd@.service  already has a
   BindsTo=sys-subsystem-net-devices-%i.device
so it will exit if the device goes away, but unfortunately
does not include
   Requisite=sys-subsystem-net-devices-%i.device

For the general case you probably are not running  hostapd
on a non permanant wifi interface, but for the timing software
I have gone with the internal wifi  wlan0  being used for
timing management - requred to manage things, though you can
do it by eth0 as well
Then  wlan1  -  as a usb plugin adapter, which can also have
bigger aerials for longer range to publish results to people
at the track.  This one is optional, and often not plugged in
when testing at home.  This will avoid filling the logs with
errors, but will start it when plugged in, and stop when removed.