Add a header to the management directory listing
Currently leaving the  html/HSV_Timing/management/  with an
accessable directory listing.  This will allow an admin to
pull up partial and development pages.

But add lighttpd support for putting a header on a directory
listing, and add a header to the management directory that
will point to the  management.html  page which presents
the main entry points for management in a better layout.