Big update to  forward_results
Capture status of the rsync, both as RC and text output.

Implement better handling of transfer errors.

If there is a permissions error, it likely means that
the results pages have been set to readonly, so don't
keep trying to update it.  Touch the transfer done
timestamp.  No further updates will be tried until there
is a data change.

If we get a 'Could not resolve hostname' then assume
our network connection is not up yet.  Do a lockout
of runs for 5 mins so we don't spam an invalid connection.