Changes to  config.php
Add a definition for  str_contains  based on str_pos
str_contains was added in PHP 8,  the older pi4b is PHP 7.4.33

The default value for booleans changed to be  "bogus"
The GUI is presenting them as a choose from a list, so one
is displayed.  When you save the first time with the new config
code, the value that was displayed will always be a change, and
thus gets saved.

When extracting values from the config, do an isset() for the full
path, before if  timing.inputs.{button|green|...}  existed, then
assumed that both  gpio  and  falling_edge  also existed.