Smarter loop for time management panels
Refactor the look to allow look behind and look ahead
so that all duplicate entries for a car can be found, not
just the second and subsequent ones.

Also allow times based on seeing the -ve of a car number
being a hidden time for the +ve version of the car.